+++ Changelog.md
@@ -1,6 +1,7 @@
 # scoop
 
 ## git master
+* [!5] Define operator<<(scoop::Color &color) out of scoop namespace
 * [!4] Fix bug in qualitative color palettes creation
 * [!3] Added sequential color map construction from palette and two boundary values.
 * [!2] Added ColorBrewer palettes (sequential, diverging and qualitative)